requirements.txt
flask==2.0.1
flask-sqlalchemy==2.5.1
flask-login==0.5.0
Werkzeug==2.0.1
pandas==1.3.3
prophet==1.0.1
matplotlib==3.4.3
plotly==5.3.1
mysqlclient==2.0.3
python-dotenv==0.19.0
Pillow==8.3.2
face-recognition==1.3.0
qrcode==7.3
numpy==1.21.2
scikit-learn==0.24.2
opencv-python==4.5.3.56
gunicorn==20.1.0

# RFID/Hardware Dependencies (optional - install manually if needed)
# bleak>=0.19.0  # For Bluetooth Low Energy RFID communication
# pyserial>=3.5  # For serial communication with hardware
flask-cors==3.0.10
bleak==0.19.5
asyncio-mqtt==0.11.1
pyserial==3.5


